+++ core/src/main/java/io/github/scroojalix/npcmanager/commands/subcommands/modifications/InteractEventModification.java
@@ -7,10 +7,10 @@ import org.bukkit.command.CommandSender;
 
 import io.github.scroojalix.npcmanager.NPCMain;
 import io.github.scroojalix.npcmanager.commands.SubCommand;
-import io.github.scroojalix.npcmanager.utils.PluginUtils;
-import io.github.scroojalix.npcmanager.utils.interactions.CommandInteraction;
-import io.github.scroojalix.npcmanager.utils.interactions.InteractionsManager;
-import io.github.scroojalix.npcmanager.utils.npc.NPCData;
+import io.github.scroojalix.npcmanager.common.PluginUtils;
+import io.github.scroojalix.npcmanager.common.interactions.CommandInteraction;
+import io.github.scroojalix.npcmanager.common.interactions.InteractionsManager;
+import io.github.scroojalix.npcmanager.common.npc.NPCData;
 
 public class InteractEventModification extends SubCommand {