+++ v1_16_R3/src/main/java/io/github/scroojalix/npcmanager/nms/v1_16_R3/EntityNMSPlayer.java
@@ -2,8 +2,8 @@ package io.github.scroojalix.npcmanager.nms.v1_16_R3;
 
 import com.mojang.authlib.GameProfile;
 
+import io.github.scroojalix.npcmanager.common.npc.NPCData;
 import io.github.scroojalix.npcmanager.nms.interfaces.NMSPlayer;
-import io.github.scroojalix.npcmanager.utils.npc.NPCData;
 import net.minecraft.server.v1_16_R3.DataWatcherRegistry;
 import net.minecraft.server.v1_16_R3.EntityPlayer;
 import net.minecraft.server.v1_16_R3.MinecraftServer;